+++ Realm/RLMRealm.mm
@@ -188,15 +188,36 @@ static NSArray *s_objectDescriptors = nil;
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
         s_defaultRealmPath = [RLMRealm writeablePathForFile:c_defaultRealmFileName];
+
+#if !TARGET_OS_IPHONE
+        [[NSFileManager defaultManager] createDirectoryAtPath:[s_defaultRealmPath stringByDeletingLastPathComponent]
+                                  withIntermediateDirectories:YES
+                                                   attributes:nil
+                                                        error:nil];
+#endif
     });
     return s_defaultRealmPath;
 }
 
 + (NSString *)writeablePathForFile:(NSString*)fileName
 {
-    NSArray *paths = NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES);
-    NSString *documentsDirectory = [paths objectAtIndex:0];
-    return [documentsDirectory stringByAppendingPathComponent:fileName];
+#if TARGET_OS_IPHONE
+    // On iOS the Documents directory isn't user-visible, so put files there
+    NSString *path = NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES)[0];
+#else
+    // On OS X it is, so put files in Application Support. If we aren't running
+    // in a sandbox, put it in a subdirectory based on the bundle identifier
+    // to avoid accidentally sharing files between applications
+    NSString *path = NSSearchPathForDirectoriesInDomains(NSApplicationSupportDirectory, NSUserDomainMask, YES)[0];
+    if (![[NSProcessInfo processInfo] environment][@"APP_SANDBOX_CONTAINER_ID"]) {
+        NSString *identifier = [[NSBundle mainBundle] bundleIdentifier];
+        if ([identifier length] == 0) {
+            identifier = [[[NSBundle mainBundle] executablePath] lastPathComponent];
+        }
+        path = [path stringByAppendingPathComponent:identifier];
+    }
+#endif
+    return [path stringByAppendingPathComponent:fileName];
 }
 
 + (instancetype)defaultRealm